# public/openapi.yaml
openapi: 3.0.3
info:
  title: Pawlink API
  version: 1.0.0
  description: |
    API backend de Pawlink — plataforma B2B SaaS para refugios de animales en LATAM.

    - **F1 Shelter Hub**: gestión de animales y solicitudes de adopción (privado, shelter)
    - **F2 Smart Adoption**: galería pública + matching de compatibilidad con IA (Groq)
    - **F3 Lost & Found**: reportes de mascotas perdidas/encontradas, geo-alertas (PostGIS) y matching por visión (AWS Rekognition)

    Nota: los endpoints aún no validan sesión (MVP). El contrato F1 exige auth de shelter — pendiente.
servers:
  - url: https://pawlink-theta.vercel.app
    description: Producción (Vercel)
  - url: http://localhost:3000
    description: Desarrollo local
...
paths:
...
  /api/adoption-requests:
    get:
      tags: [Adoption Requests (F1/F2)]
      summary: Lista las solicitudes de adopción de un shelter
      parameters:
        - { name: shelter_id, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: string, enum: [pending, seen, approved, rejected] } }
      responses:
        '200':
          description: Solicitudes con datos del animal y la familia
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests:
                    type: array
                    items: { $ref: '#/components/schemas/AdoptionRequest' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      tags: [Adoption Requests (F1/F2)]
      summary: Crea una solicitud de adopción (sin cuenta — contacto inline)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [animal_id, shelter_id, full_name, email]
              properties:
                animal_id: { type: string, format: uuid }
                shelter_id: { type: string, format: uuid }
                full_name: { type: string, maxLength: 120, example: Ana García }
                email: { type: string, format: email, maxLength: 255, description: Se guarda en minúsculas — clave de dedupe junto con animal_id }
                phone: { type: string, maxLength: 20, example: '+52 55 1234 5678' }
                family_profile:
                  allOf: [{ $ref: '#/components/schemas/FamilyProfile' }]
                  description: Cuestionario anidado (shape del contrato F2), opcional.
                compatibility_score:
                  type: number
                  minimum: 0
                  maximum: 100
                  description: Score calculado por /api/matching — se guarda para que el shelter lo vea en su dashboard
                  example: 94.5
                compatibility_reasons:
                  type: array
                  items: { type: string }
                  example: [Temperamento tranquilo, Buena con niños]
      responses:
        '201':
          description: Solicitud creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  request:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      status: { type: string, example: pending }
                      animal_id: { type: string, format: uuid }
                      shelter_id: { type: string, format: uuid }
                      created_at: { type: string, format: date-time }
                  message: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: Ya existe una solicitud pendiente para este animal con el mismo email
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  responses:
    BadRequest:
      description: Parámetros faltantes o inválidos
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServerError:
      description: Error interno (mensaje de Supabase)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: shelter_id is required }
...
    FamilyProfile:
      type: object
      required: [living_space, lifestyle, experience]
      properties:
        living_space: { type: string, enum: [apartment, house_no_yard, house_yard] }
        lifestyle: { type: string, enum: [sedentary, moderate, active] }
        experience: { type: string, enum: [none, some, experienced] }
        has_children: { type: boolean, default: false }
        has_other_pets: { type: boolean, default: false }
    AdoptionRequest:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [pending, seen, approved, rejected] }
        compatibility_score: { type: number, nullable: true, example: 94.5 }
        compatibility_reasons:
          type: array
          items: { type: string }
        notes: { type: string, nullable: true }
        animal:
          type: object
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            photo_urls: { type: array, items: { type: string, format: uri } }
        family:
          type: object
          description: Contacto y cuestionario guardados en la propia solicitud (no hay tabla de familias)
          properties:
            full_name: { type: string }
            email: { type: string, format: email }
            phone: { type: string, nullable: true }
            living_space: { type: string }
            has_children: { type: boolean }
            has_other_pets: { type: boolean }
        created_at: { type: string, format: date-time }
...